pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "twinkl"
version = "0.1.0"
description = "Voice-first alignment journaling capstone project."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel==7.1.0",
    "jinja2==3.1.6",
    "openai==2.11.0",
    "pydantic==2.12.5",
    "python-dotenv==1.2.1",
    "pyyaml==6.0.3",
    "polars==1.36.1",
    "shiny>=1.2.0",
    "scikit-learn>=1.3.0",
    "statsmodels>=0.14.0",
    # VIF Critic training dependencies
    "torch>=2.0.0",
    "bayesian-torch>=0.5.0",
    "sentence-transformers>=2.2.0",
    "scipy>=1.10.0",
    "matplotlib>=3.7.0",
    "pytorch-tcn>=1.2.3",
    "coral-pytorch>=1.4.0",
    "ipywidgets>=8.1.8",
    "einops>=0.8.2",
]

[tool.setuptools]
# This project currently has no installable Python packages.
# Explicitly disable auto-discovery so top-level folders like
# `config/` and `logs/` don't get treated as packages.
packages = []

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = []

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
